+++ SHFB/SHFB.shfbproj
@@ -116,7 +116,8 @@
 <NamespaceSummaryItem name="Koopman.CheckPoint.SimpleGatewaySettings" isDocumented="True">CheckPoint.NET Simple Gateway Settings Classes</NamespaceSummaryItem>
 <NamespaceSummaryItem name="Koopman.CheckPoint.FastUpdate" isDocumented="True">Allows for editing existing objects without finding them first.</NamespaceSummaryItem>
 <NamespaceSummaryItem name="Koopman.CheckPoint.AccessRules" isDocumented="True">CheckPoint.NET Access Rule Classes</NamespaceSummaryItem>
-<NamespaceSummaryItem name="Koopman.CheckPoint.Special" isDocumented="False" /></NamespaceSummaries>
+<NamespaceSummaryItem name="Koopman.CheckPoint.Special" isDocumented="False" />
+<NamespaceSummaryItem name="Koopman.CheckPoint.IA" isDocumented="True">Check Point Identity Awareness Gateway API classes.</NamespaceSummaryItem></NamespaceSummaries>
     <ProjectSummary>CheckPoint.NET Namespaces</ProjectSummary>
     <SourceCodeBasePath>..\src\</SourceCodeBasePath>
     <CatalogProductId>VS</CatalogProductId>
